Average with cutoff returns now float32 image

+++ pyFAI/average.py
@@ -365,7 +365,7 @@ def average_dark(lstimg, center_method="mean", cutoff=None, quantiles=(0.5, 0.5)
         mask = ((abs(stack - center) / std) > cutoff)
         stack[numpy.where(mask)] = 0.0
         summed = stack.sum(axis=0)
-        output = summed / numpy.maximum(1, (length - mask.sum(axis=0)))
+        output = summed / numpy.float32(numpy.maximum(1, (length - mask.sum(axis=0))))
     return output
